+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.2.1] - 2026-07-22
+
+### Fixed
+
+- Fixed haplotype network PNG export being cropped and low-resolution on Retina/HiDPI displays.
+- Fixed haplotype network edge lines being invisible in dark theme (previously hardcoded to black).
+
+---
+
 ## [1.2.0] - 2026-07-19
 
 ### Added